+++ src/Table.fs
@@ -2,34 +2,22 @@ namespace MF.ConsoleStyle
 
 [<RequireQualifiedAccess>]
 module private Table =
-    let private tee f a =
-        f a
-        a
+    open Words
+    open WordLength.Operators
 
-    type private WordLength = WordLength of int
-
-    [<AutoOpen>]
-    module private WordLengthOperators =
-        let inline (<+>) (WordLength a) (WordLength b) = WordLength (a + b)
-        let inline (<->) (WordLength a) (WordLength b) = WordLength (a - b)
-
-    type private ColumnIndex = ColumnIndex of int
-
-    type private MaxWordLengths = MaxWordLengths of Map<ColumnIndex, WordLength>
-
-    type private Column = Column of string
+    type private Column = Column of Word
 
     [<RequireQualifiedAccess>]
     module private Column =
-        let value (Column column) = column
-        let length removeMarkup (Column column) = WordLength (column |> removeMarkup |> String.length)
+        let word (Column word) = word
+        let value = word >> Word.value
 
     type private RowLine = RowLine of string
 
     [<RequireQualifiedAccess>]
     module private RowLine =
         let ofColumns columns = columns |> List.map Column.value |> String.concat " " |> RowLine
-        let toColumns (RowLine line) = line.Split " " |> Array.toList |> List.map Column
+        let toColumns (RowLine line) = line.Split " " |> Array.toList |> List.map (Word >> Column)
 
         let value (RowLine line) = line
 
@@ -47,26 +35,15 @@ module private Table =
             | Line line -> line |> RowLine.toColumns
             | Columns columns -> columns
 
-        let columnsLengths removeMarkup = columns >> List.mapi (fun columnIndex column -> (ColumnIndex columnIndex, column |> Column.length removeMarkup))
+        let toWords = columns >> List.map Column.word
 
-        let format removeMarkup (MaxWordLengths maxWordLengths) = function
+        let format removeMarkup maxWordLengths = function
             | Line line -> line
             | Columns columns ->
                 columns
-                |> List.mapi (fun index (Column word) ->
-                    let (WordLength columnLength) = maxWordLengths |> Map.find (ColumnIndex index)
-
-                    let markupLength =
-                        let lengthWithoutMarkup = word |> removeMarkup |> String.length
-                        word.Length - lengthWithoutMarkup
-
-                    let realColumnLength =
-                        match markupLength with
-                        | markupLength when markupLength > 0 -> columnLength + markupLength
-                        | _ -> columnLength
-
-                    Column (sprintf " %-*s" (realColumnLength - 1) word)
-                )
+                |> List.map Column.word
+                |> Line.format removeMarkup (fun realColumnLength -> sprintf " %-*s" (realColumnLength - 1)) maxWordLengths
+                |> List.map Column
                 |> RowLine.ofColumns
 
     [<RequireQualifiedAccess>]
@@ -94,6 +71,8 @@ module private Table =
             | Rows.Empty -> []
             | Rows.Rows rows -> rows
 
+        let toLines = rows >> List.map Row.toWords
+
         let render renderSeparatorRow renderRow = function
             | Rows.Empty -> ()
             | Rows.Rows rows ->
@@ -110,6 +89,7 @@ module private Table =
                 separator
                 |> string
                 |> String.replicate wordLength
+                |> Word
                 |> Column
 
             wordLengths
@@ -119,33 +99,13 @@ module private Table =
             |> Row.line
             |> renderLine
 
-    [<RequireQualifiedAccess>]
-    module private MaxWordLengths =
-        let perColumnsInRows removeMarkup rows =
-            let rec findMaxLengths maxLengths = function
-                | [] ->
-                    maxLengths
-                    |> Map.map (fun _ length -> length <+> WordLength 2)    // +1 space for both sides of the word ` word `
-                    |> MaxWordLengths
-                | (columnIndex, wordLength) :: rows ->
-                    let maxLengthForColumn =
-                        match maxLengths |> Map.tryFind columnIndex with
-                        | Some currentMaxWordLength -> max currentMaxWordLength wordLength
-                        | _ -> wordLength
-
-                    rows |> findMaxLengths (maxLengths.Add (columnIndex, maxLengthForColumn))
-
-            rows
-            |> Rows.rows
-            |> List.collect (Row.columnsLengths removeMarkup)
-            |> findMaxLengths Map.empty
-
     let render
-        (removeMarkup: string -> string)
         (renderHeaderLine: string -> unit)
         (renderRowLine: string -> unit)
         (header: string list)
         (rows: (string list) list) =
+
+        let removeMarkup = Render.Markup.removeMarkup
 
         let formatRow maxWordLengths = Row.format removeMarkup maxWordLengths >> RowLine.value
 
@@ -157,14 +117,18 @@ module private Table =
             Separator '-'
             |> Separator.renderRow renderLine wordLengths
 
-        let getMaxWordLengthsPerColumn = MaxWordLengths.perColumnsInRows removeMarkup
+        let getMaxWordLengthsPerColumn rows =
+            rows
+            |> Rows.toLines
+            |> MaxWordLengths.perWordsInLine removeMarkup
+            |> MaxWordLengths.map ((<+>) (WordLength 2))    // +1 space for both sides of the word ` word `
 
         let header =
             match header with
             | [] -> Header.Empty
             | words ->
                 words
-                |> List.map Column
+                |> List.map (removeMarkup >> Word >> Column)
                 |> Columns
                 |> Header.Row
 
@@ -173,7 +137,7 @@ module private Table =
             | [] -> Rows.Empty
             | rows ->
                 rows
-                |> List.map (List.map Column >> Columns)
+                |> List.map (List.map (Word >> Column) >> Columns)
                 |> Rows.Rows
 
         let wordMaxLengths =